refactor: extract operational and product read registrars (#1731)
