+++ control_plane/http_routes/__init__.py
@@ -1,9 +1,37 @@
 from control_plane.http_routes.ingress import register_ingress_read_routes
+from control_plane.http_routes.operational_records import (
+    register_deployment_promotion_read_routes,
+    register_inventory_operation_read_routes,
+    register_managed_secret_read_routes,
+)
+from control_plane.http_routes.products import (
+    ProductReadRouteDependencies,
+    product_profile_context_cutover_contexts_allowed,
+    register_agent_context_read_routes,
+    register_product_config_status_read_routes,
+    register_product_context_audit_read_routes,
+    register_product_environment_read_routes,
+    register_product_profile_read_routes,
+    register_protected_artifact_read_routes,
+    require_product_profile_read_store,
+)
 from control_plane.http_routes.support import ReadRouteDependencies
 from control_plane.http_routes.topology import register_topology_read_routes
 
 __all__ = (
+    "ProductReadRouteDependencies",
     "ReadRouteDependencies",
+    "product_profile_context_cutover_contexts_allowed",
+    "register_agent_context_read_routes",
+    "register_deployment_promotion_read_routes",
     "register_ingress_read_routes",
+    "register_inventory_operation_read_routes",
+    "register_managed_secret_read_routes",
+    "register_product_config_status_read_routes",
+    "register_product_context_audit_read_routes",
+    "register_product_environment_read_routes",
+    "register_product_profile_read_routes",
+    "register_protected_artifact_read_routes",
     "register_topology_read_routes",
+    "require_product_profile_read_store",
 )
