+++ docs/service-boundary.md
@@ -385,12 +385,16 @@ New or changed service route families must preserve the completed HTTP boundary:
 typed Pydantic response, and focused OpenAPI assertions. Use it as the small
 contract shape for future route-family slices.
 
-Read-only ingress and topology route families are registered from
-dependency-explicit modules under `control_plane.http_routes`. `http_app.py`
-remains the composition root: it passes frozen dependency objects and callables
-into the registrars, and each registrar calls the Launchplane FastAPI app's
-custom `add_api_route` directly so shared error-schema handling and route order
-remain centralized without an `APIRouter` compatibility layer.
+Read-only ingress, topology, operational-record, managed-secret, and product
+route families are registered from dependency-explicit modules under
+`control_plane.http_routes`. `http_app.py` remains the composition root: it
+passes frozen dependency objects and callables into the registrars, and each
+registrar calls the Launchplane FastAPI app's custom `add_api_route` directly so
+shared error-schema handling and route order remain centralized without an
+`APIRouter` compatibility layer. Route families with intervening write or UI
+registrations expose multiple registrar entrypoints rather than being reordered;
+in particular, product environment config-status remains the final API route
+registered immediately before the exception handlers.
 
 Frontend contract generation uses the same boundary. Run
 `uv run launchplane service export-openapi --output frontend/generated/openapi-canonical.json`
