+++ tests/http_app_test_support.py
@@ -2773,41 +2773,6 @@ def _local_operator_bearer_config(*, token_label: str = "local-owner-read") -> B
 type _AsgiResponse = Response
 
 
-async def _get_config_status(
-    app: FastAPI,
-    *,
-    product: str = "example-site",
-    environment: str = "prod",
-    authorization: str = "Bearer valid-token",
-) -> _AsgiResponse:
-    headers = {"Authorization": authorization} if authorization else {}
-    return await _asgi_get(
-        app,
-        f"/v1/products/{product}/environments/{environment}/config-status",
-        headers=headers,
-    )
-
-
-async def _get_repo_product_mapping(
-    app: FastAPI,
-    *,
-    authorization: str = "Bearer valid-token",
-) -> _AsgiResponse:
-    headers = {"Authorization": authorization} if authorization else {}
-    return await _asgi_get(app, "/v1/repo-product-mapping", headers=headers)
-
-
-async def _get_agent_context(
-    app: FastAPI,
-    *,
-    repository: str = "",
-    authorization: str = "Bearer valid-token",
-) -> _AsgiResponse:
-    headers = {"Authorization": authorization} if authorization else {}
-    suffix = f"?{urlencode({'repository': repository})}" if repository else ""
-    return await _asgi_get(app, f"/v1/agent/context{suffix}", headers=headers)
-
-
 async def _get_work_graph_snapshot(
     app: FastAPI,
     *,
@@ -3773,85 +3738,6 @@ async def _post_odoo_prod_promotion_run(
     )
 
 
-async def _get_products(
-    app: FastAPI,
-    *,
-    authorization: str = "Bearer valid-token",
-) -> _AsgiResponse:
-    headers = {"Authorization": authorization} if authorization else {}
-    return await _asgi_get(app, "/v1/products", headers=headers)
-
-
-async def _get_product(
-    app: FastAPI,
-    product: str = "example-site",
-    *,
-    authorization: str = "Bearer valid-token",
-) -> _AsgiResponse:
-    headers = {"Authorization": authorization} if authorization else {}
-    return await _asgi_get(app, f"/v1/products/{product}", headers=headers)
-
-
-async def _get_product_activity(
-    app: FastAPI,
-    product: str = "example-site",
-    *,
-    authorization: str = "Bearer valid-token",
-) -> _AsgiResponse:
-    headers = {"Authorization": authorization} if authorization else {}
-    return await _asgi_get(app, f"/v1/products/{product}/activity", headers=headers)
-
-
-async def _get_product_environments(
-    app: FastAPI,
-    product: str = "example-site",
-    *,
-    authorization: str = "Bearer valid-token",
-) -> _AsgiResponse:
-    headers = {"Authorization": authorization} if authorization else {}
-    return await _asgi_get(app, f"/v1/products/{product}/environments", headers=headers)
-
-
-async def _get_product_environment(
-    app: FastAPI,
-    product: str = "example-site",
-    environment: str = "prod",
-    *,
-    authorization: str = "Bearer valid-token",
-) -> _AsgiResponse:
-    headers = {"Authorization": authorization} if authorization else {}
-    return await _asgi_get(
-        app,
-        f"/v1/products/{product}/environments/{environment}",
-        headers=headers,
-    )
-
-
-async def _get_protected_artifacts(
-    app: FastAPI,
-    *,
-    product: str,
-    context: str = "",
-    authorization: str = "Bearer local-operator-token",
-    headers: dict[str, str] | None = None,
-) -> _AsgiResponse:
-    params: dict[str, str] = {}
-    if product:
-        params["product"] = product
-    if context:
-        params["context"] = context
-    query_string = urlencode(params)
-    suffix = f"?{query_string}" if query_string else ""
-    request_headers = dict(headers or {})
-    if authorization:
-        request_headers["Authorization"] = authorization
-    return await _asgi_get(
-        app,
-        f"/v1/artifacts/protected{suffix}",
-        headers=request_headers,
-    )
-
-
 async def _get_driver_descriptors(
     app: FastAPI,
     *,
@@ -3955,98 +3841,6 @@ def _ingress_canary_route_apply_payload(
         "canary_key": canary_key,
         "reason": reason,
     }
-
-
-async def _get_deployment_record(
-    app: FastAPI,
-    record_id: str,
-    *,
-    authorization: str = "Bearer valid-token",
-    headers: dict[str, str] | None = None,
-) -> _AsgiResponse:
-    request_headers = dict(headers or {})
-    if authorization:
-        request_headers["Authorization"] = authorization
-    return await _asgi_get(app, f"/v1/deployments/{record_id}", headers=request_headers)
-
-
-async def _get_promotion_record(
-    app: FastAPI,
-    record_id: str,
-    *,
-    authorization: str = "Bearer valid-token",
-    headers: dict[str, str] | None = None,
-) -> _AsgiResponse:
-    request_headers = dict(headers or {})
-    if authorization:
-        request_headers["Authorization"] = authorization
-    return await _asgi_get(app, f"/v1/promotions/{record_id}", headers=request_headers)
-
-
-async def _get_environment_inventory(
-    app: FastAPI,
-    context: str,
-    instance: str,
-    *,
-    authorization: str = "Bearer valid-token",
-    headers: dict[str, str] | None = None,
-) -> _AsgiResponse:
-    request_headers = dict(headers or {})
-    if authorization:
-        request_headers["Authorization"] = authorization
-    return await _asgi_get(
-        app,
-        f"/v1/inventory/{context}/{instance}",
-        headers=request_headers,
-    )
-
-
-async def _get_recent_operations(
-    app: FastAPI,
-    context: str,
-    *,
-    authorization: str = "Bearer valid-token",
-    headers: dict[str, str] | None = None,
-) -> _AsgiResponse:
-    request_headers = dict(headers or {})
-    if authorization:
-        request_headers["Authorization"] = authorization
-    return await _asgi_get(
-        app,
-        f"/v1/contexts/{context}/operations/recent",
-        headers=request_headers,
-    )
-
-
-async def _get_product_profiles(
-    app: FastAPI,
-    *,
-    driver_id: str = "",
-    authorization: str = "Bearer valid-token",
-    headers: dict[str, str] | None = None,
-) -> _AsgiResponse:
-    request_headers = dict(headers or {})
-    if authorization:
-        request_headers["Authorization"] = authorization
-    suffix = f"?{urlencode({'driver_id': driver_id})}" if driver_id else ""
-    return await _asgi_get(app, f"/v1/product-profiles{suffix}", headers=request_headers)
-
-
-async def _get_product_profile(
-    app: FastAPI,
-    product: str = "sellyouroutboard",
-    *,
-    authorization: str = "Bearer valid-token",
-    headers: dict[str, str] | None = None,
-) -> _AsgiResponse:
-    request_headers = dict(headers or {})
-    if authorization:
-        request_headers["Authorization"] = authorization
-    return await _asgi_get(
-        app,
-        f"/v1/product-profiles/{product}",
-        headers=request_headers,
-    )
 
 
 async def _post_product_profile(
@@ -4251,80 +4045,6 @@ async def _post_legacy_context_cleanup_apply(
         payload=payload,
         raw_body=raw_body,
     )
-
-
-async def _get_context_cutover_audit(
-    app: FastAPI,
-    *,
-    product: str = "sellyouroutboard",
-    source_context: str = "sellyouroutboard-testing",
-    target_context: str = "sellyouroutboard",
-    preview_context: str = "",
-    authorization: str = "Bearer valid-token",
-    headers: dict[str, str] | None = None,
-) -> _AsgiResponse:
-    request_headers = dict(headers or {})
-    if authorization:
-        request_headers["Authorization"] = authorization
-    params = {
-        "source_context": source_context,
-        "target_context": target_context,
-    }
-    if preview_context:
-        params["preview_context"] = preview_context
-    return await _asgi_get(
-        app,
-        f"/v1/product-profiles/{product}/context-cutover-audit?{urlencode(params)}",
-        headers=request_headers,
-    )
-
-
-async def _get_context_secret_statuses(
-    app: FastAPI,
-    context: str,
-    *,
-    authorization: str = "Bearer valid-token",
-    headers: dict[str, str] | None = None,
-) -> _AsgiResponse:
-    request_headers = dict(headers or {})
-    if authorization:
-        request_headers["Authorization"] = authorization
-    return await _asgi_get(
-        app,
-        f"/v1/contexts/{context}/secrets",
-        headers=request_headers,
-    )
-
-
-async def _get_instance_secret_statuses(
-    app: FastAPI,
-    context: str,
-    instance: str,
-    *,
-    authorization: str = "Bearer valid-token",
-    headers: dict[str, str] | None = None,
-) -> _AsgiResponse:
-    request_headers = dict(headers or {})
-    if authorization:
-        request_headers["Authorization"] = authorization
-    return await _asgi_get(
-        app,
-        f"/v1/contexts/{context}/instances/{instance}/secrets",
-        headers=request_headers,
-    )
-
-
-async def _get_secret_status(
-    app: FastAPI,
-    secret_id: str,
-    *,
-    authorization: str = "Bearer valid-token",
-    headers: dict[str, str] | None = None,
-) -> _AsgiResponse:
-    request_headers = dict(headers or {})
-    if authorization:
-        request_headers["Authorization"] = authorization
-    return await _asgi_get(app, f"/v1/secrets/{secret_id}", headers=request_headers)
 
 
 async def _get_preview_record(
@@ -4753,95 +4473,6 @@ class _EmptyStore:
 
     def close(self) -> None:
         return None
-
-
-class _SecretStatusProbeStore:
-    def __init__(self) -> None:
-        self.calls: list[str] = []
-
-    def read_secret_record(self, secret_id: str) -> object:
-        self.calls.append(f"read_secret_record:{secret_id}")
-        raise FileNotFoundError(secret_id)
-
-    def list_secret_records(
-        self,
-        *,
-        integration: str = "",
-        context_name: str = "",
-        instance_name: str = "",
-        limit: int | None = None,
-    ) -> tuple[object, ...]:
-        del integration, context_name, instance_name, limit
-        self.calls.append("list_secret_records")
-        return ()
-
-    def read_secret_version(self, version_id: str) -> object:
-        self.calls.append(f"read_secret_version:{version_id}")
-        raise FileNotFoundError(version_id)
-
-    def list_secret_versions(self, *, secret_id: str) -> tuple[object, ...]:
-        self.calls.append(f"list_secret_versions:{secret_id}")
-        return ()
-
-    def list_secret_bindings(
-        self,
-        *,
-        integration: str = "",
-        context_name: str = "",
-        instance_name: str = "",
-        limit: int | None = None,
-    ) -> tuple[object, ...]:
-        del integration, context_name, instance_name, limit
-        self.calls.append("list_secret_bindings")
-        return ()
-
-    def list_secret_audit_events(self, *, secret_id: str) -> tuple[object, ...]:
-        self.calls.append(f"list_secret_audit_events:{secret_id}")
-        return ()
-
-
-class _RecentOperationsProbeStore:
-    def __init__(self) -> None:
-        self.calls: list[str] = []
-
-    def list_environment_inventory(self) -> tuple[EnvironmentInventory, ...]:
-        self.calls.append("list_environment_inventory")
-        return ()
-
-    def list_deployment_records(
-        self,
-        *,
-        context_name: str = "",
-        instance_name: str = "",
-        limit: int | None = None,
-    ) -> tuple[DeploymentRecord, ...]:
-        del context_name, instance_name, limit
-        self.calls.append("list_deployment_records")
-        return ()
-
-    def list_promotion_records(
-        self,
-        *,
-        context_name: str = "",
-        from_instance_name: str = "",
-        to_instance_name: str = "",
-        limit: int | None = None,
-    ) -> tuple[PromotionRecord, ...]:
-        del context_name, from_instance_name, to_instance_name, limit
-        self.calls.append("list_promotion_records")
-        return ()
-
-    def list_preview_records(
-        self,
-        *,
-        context_name: str = "",
-        anchor_repo: str = "",
-        anchor_pr_number: int | None = None,
-        limit: int | None = None,
-    ) -> tuple[PreviewRecord, ...]:
-        del context_name, anchor_repo, anchor_pr_number, limit
-        self.calls.append("list_preview_records")
-        return ()
 
 
 class _PreviewRecordOnlyStore:
